Add named socket example in Python
The server creates a domain socket as a file. It then listens to this
socket for any clients. When a client message arrives, it performs an
action and then returns a result. Then it continues listening. So far, I
haven't encountered any busy waiting issues.

The main difference between named pipes, fifos, and sockets is that
sockets are two-way. The performans is reasonably similar, with fifos
perhaps being a little faster.

There are many many different types of sockets. I am currently using
domain sockets. These behave as files on the filesystem and can be used
for very fast local communcation between processes. This differs from a
TCP socket which would have an ip address (e.g., at the localhost) and a
port. TCP allows communication across a network, but is slower.

Python has a great sockets library. R, however, has nothing at all, so I
will have to build one through the C API.